Create mark2.html
If you do want to attribute my assistance, you can simply mention that the code was developed with the help of ChatGPT, a language model trained by OpenAI. You can also provide a link to the OpenAI website (https://openai.com/) if you'd like. However, this is entirely optional and not necessary.